% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supisd.R
\name{supisd}
\alias{supisd}
\title{Distributional/Tail Structural Change Test}
\usage{
supisd(
  data,
  breakstart = NULL,
  breakend = NULL,
  breaklength = NULL,
  evalrange.l = NULL,
  evalrange.r = NULL,
  width = NULL,
  S = NULL
)
}
\arguments{
\item{data}{data in chronological order to be tested.}

\item{breakstart}{a numeric value giving the starting period of all possible
breaks. Default leaves 20 observations from the start of the data.}

\item{breakend}{a numeric value giving the ending period of all possible
breaks. Default leaves 20 observations from the end of the data.}

\item{breaklength}{a numeric value giving the length between break points.
Default is 1.}

\item{evalrange.l}{lower bound of the evaluation support. Either a numeric
value from 0 to 1 indicating the percentage of data range from the
minimum of the data, or a character sting "mean" indicating the mean
of the data. Default is 10 percent of data range to the left of the
minimum of the data.}

\item{evalrange.r}{upper bound of the evaluation support. Either a numeric
value from 0 to 1 indicating the percentage of data range from the
minimum of the data, or a character sting "mean" indicating the mean
of the data. Default is 10\% of the data range to the right of the
maximum of the data.}

\item{width}{space between the points in the grid for kernel estimation.
Default results in 500 points at which the density is estimated.}

\item{S}{number of simulations for generating the null distribution of the
test statistic. Default is 1000.}
}
\value{
test statistic, estimated break and p value
}
\description{
The function supisd() performs the distributional/tail structural change test
proposed in Lu and Ker (2024).
Note: This version is for test only. Please contact the authors if have any
questions. Thank you.
}
\details{
A Sup-type nonparametric test for structural change with an unknown break
in time-series data as proposed in Lu and Ker (2024). The test has power
against structural change in any moment and can be tailored to a specific
range of the underlying distribution. The function uses Gaussian kernel
function and leave-one-out cross validation for finding the optimal bandwidth
for kernel estimation. Test p value is found via randomization method as
detailed in Lu and Ker (2024). It is suggested by simulations that the
possible breaks to be tested should not be too close to the start/end of
data (leaving at least 20 observations from the start/end is recommended).
}
\examples{
#width and S in examples below are set to reduce running time.
#Still, it may take a few seconds.
data <- rnorm(100)
#Tests structural change over the entire domain of data
supisd(data, width=0.1, S=200)
#Test structural change within the lower tail (below mean) of data
supisd(data, evalrange.r="mean", width=0.1, S=200)
#Test structural change within the upper tail (above mean) of data
supisd(data, evalrange.l="mean", width=0.1, S=200)

}
\references{
Hanjun Lu, Alan P Ker, Testing for distributional structural
change with unknown breaks: application to pricing crop insurance contracts,
Journal of the Royal Statistical Society Series C: Applied Statistics, 2024;,
qlae020, https://doi.org/10.1093/jrsssc/qlae020
}
\author{
Hanjun Lu and Alan Ker
}
